requirements.txt
boto3==1.20.24
gql==3.2.0
joblib==1.1.0
neo4j==4.4.5
pandas==1.3.4
python-dotenv==0.20.0
requests==2.26.0
s3fs==2022.2.0
tqdm==4.62.3
